Don't use fill-column-indicator or diff-hl
This experiment sort of failed.

First of all, diff-hl clashes with magit and causes emacs to hang
up. Second of all, fci was *not* the cause of my ox-rss export problems,
so I left it :ensure'd but not enabled. I commented out the advice
because it won't be turned on via a hook.